--- Controls-demo/list_new/Grouped/Nested/Nested.wml
@@ -1,22 +1,0 @@
-<div class="controlsDemo__wrapper">
-    <Controls.scroll:Container>
-        <Controls.list:View
-                keyProperty="key"
-                source="{{_rootSource}}"
-                groupProperty="store">
-            <ws:groupTemplate>
-                <ws:partial template="Controls/list:GroupTemplate" fontSize="xl"/>
-            </ws:groupTemplate>
-            <ws:itemTemplate>
-                <ws:partial template="Controls/list:ItemTemplate">
-                    <ws:contentTemplate>
-                        <Controls.list:View
-                                keyProperty="key"
-                                source="{{_nestedSource}}"
-                                groupProperty="brand"/>
-                    </ws:contentTemplate>
-                </ws:partial>
-            </ws:itemTemplate>
-        </Controls.list:View>
-    </Controls.scroll:Container>
-</div>
